requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
Django==3.0.3
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
dj-database-url==0.4.2
gunicorn==19.7.1
psycopg2==2.8.4
whitenoise==3.3.1